# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Consultas de transacciones",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/backend-api-consultas/public"
        },
        {
            "url": "http://localhost/backend-api-consultas/public"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "TOKEN",
                "description": "User Login Token",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Logged In Successfully."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|0JJCOd0QSTajchCUf5fYPex0ZKThSjLWefe3OfTb"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Login Unsuccessfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/consulta-pagos": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Consulta de transacciones",
                "description": "Consulta de transacciones pagadas por entidad",
                "operationId": "72aa7439e21ddbffa4aed650bb5e7cab",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registros encontrados"
                                        },
                                        "datos": {
                                            "type": "array",
                                            "items": {},
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/verifica-pagos": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Verifica pagos",
                "description": "Verificacion de transacciones consultadas",
                "operationId": "verifica_pagos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_transaccion_motor": {
                                        "type": "object",
                                        "collectionFormat": "multi",
                                        "example": "[2000000000]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": 202
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Guardado exitoso."
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unsuccessfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": 400
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "id_transaccion_motor requerido."
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consulta-folios": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Consulta de folios",
                "description": "consulta de folios",
                "operationId": "consulta_folios",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_transaccion_motor": {
                                        "type": "object",
                                        "collectionFormat": "multi",
                                        "example": "[2000000000]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": 202
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Registros encontrados"
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unsuccessfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": 400
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Sin Registros encontrado."
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consulta-archivos": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Consulta archivos",
                "description": "consulta por fecha y genera una url de descarga de archivo en txt",
                "operationId": "consulta_archivos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha_registro": {
                                        "type": "object",
                                        "collectionFormat": "multi",
                                        "example": "2021-02-22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Archivo disponible, descarga una vez, expira en 24hrs."
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://website.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "in": "header",
                "bearerAuth": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}